==> 1. Cooler Box Data Extraction from API.ipynb <==
# Import Libraries for data extraction through API
import requests
import json
import pandas as pd
from pandas import json_normalize
from datetime import datetime, timedelta

# Mentioning the API key from where data has been extracted
API_KEY={''}

date=datetime.strptime(input("Enter Date('%Y-%m-%d'):"), '%Y-%m-%d')

# 1.1 Data Extraction- Monthly-Parameter- Light

# Code for Extraction of Data through API

# The size of each step in days
dataframeLight=pd.DataFrame()
day_delta = timedelta(days=1)
delta = timedelta(days=7)
print(day_delta)
print(delta)

start_date = date

tagid= [2031010108,2031010506,2031010170]
        
index= 0

while index < len(tagid):
    con_tagid = str(tagid[index])
    print(type(con_tagid),'tagid:',con_tagid)
    
    param=str('light') 
    print('Parameter Name:',param)

    start_date1 = start_date
    
    #start_date = input("Enter Date:")
    end_date = start_date1 + 31*day_delta
    print(end_date)
    Dategap=((end_date - start_date1).days)/7
    print("Dategap:",Dategap)

    for i in range(int(Dategap)):
        print ("Loop:",i)
        Y= start_date1
        X=(start_date1 + 6*day_delta)
        firstDate_time= Y.strftime("%Y-%m-%dT%H:%M:%S")
        LastDate_time= X.strftime("%Y-%m-%dT%H:%M:%S")
        print('FirstDate_time:',firstDate_time)
        print('LastDate_time:',LastDate_time)

        url=''+param+'?startTime='+firstDate_time+'&endTime='+LastDate_time+'&tagId='+con_tagid
        r=requests.post(url,headers=API_KEY)
        data = r.json()
        #print(data)
        works_data=json_normalize(data=data['data'],record_path=['telData'])
        df1=pd.DataFrame.from_dict(works_data)
        dataframeLight=pd.concat([df1,dataframeLight],ignore_index=True)
        print(dataframeLight)
       
        start_date1 += delta
        
    index = index + 1

dataframeLight['tag_id'].value_counts()

dataframeLight.to_excel('Monthly_Rawdata_Light.xlsx',index=False)

# 1.2 Data Extraction- Monthly-Parameter- Temp

# Code for Extraction of Data through API

# Creating a blank dataframe
dataframeTemp=pd.DataFrame()

# The size of each step in days
day_delta = timedelta(days=1)
delta = timedelta(days=7)
print(day_delta)
print(delta)

start_date=date

# Mentioning the sensorids

tagid= [2031010108,2031010506,2031010170]

index= 0

while index < len(tagid):
    con_tagid = str(tagid[index])
    print(type(con_tagid),'tagid:',con_tagid)
    
    param=str('temperature') 
    print('Parameter Name:',param)
    
    start_date1=start_date

    #mentioning the start date and for no of days data is required
    end_date = start_date1 + 31*day_delta      #End date is 31st days of starting day
    print('end_date:',end_date)
    Dategap=((end_date - start_date1).days)/7
    print("Dategap:",Dategap)

    for i in range(int(Dategap)):
        print ("Loop:",i)
        Y= start_date1
        X=(start_date1 + 6*day_delta)
        firstDate_time= Y.strftime("%Y-%m-%dT%H:%M:%S")
        LastDate_time= X.strftime("%Y-%m-%dT%H:%M:%S")
        print('FirstDate_time:',firstDate_time)
        print('LastDate_time:',LastDate_time)

        
        #Mentioning the URl for data extraction
        url=''+param+'?startTime='+firstDate_time+'&endTime='+LastDate_time+'&tagId='+con_tagid
        r=requests.post(url,headers=API_KEY)
        data = r.json()
       
        #Normalizing the data
        works_data=json_normalize(data=data['data'],record_path=['telData'])
        
        #appending the data into one dataframe
        df1=pd.DataFrame.from_dict(works_data)
        dataframeTemp=pd.concat([df1,dataframeTemp],ignore_index=True)
        print(dataframeTemp)
        
        start_date1 += delta

    index = index + 1

dataframeTemp.columns

dataframeTemp['tag_id'].value_counts()

dataframeTemp.to_excel('Monthly_Rawdata_Temp.xlsx',index=False)

# 1.3 Data Extraction- Monthly-Parameter- Jerk

# Code for Extraction of Data through API

# Creating a blank dataframe
dataframeJerk=pd.DataFrame()

# The size of each step in days
day_delta = timedelta(days=1)
delta = timedelta(days=7)
print(day_delta)
print(delta)

start_date=date

# Mentioning the sensorids

tagid= [2031010108,2031010506,2031010170]

index= 0

while index < len(tagid):
    con_tagid = str(tagid[index])
    print(type(con_tagid),'tagid:',con_tagid)
    
    param=str('jerks') 
    print('Parameter Name:',param)
    
    start_date1 = start_date

    #mentioning the no of days data is required
    end_date = start_date1 + 31*day_delta #End date is 31st days of starting day
    print('end_date:',end_date)
    Dategap=((end_date - start_date1).days)/7
    print("Dategap:",Dategap)

    for i in range(int(Dategap)):
        print ("Loop:",i)
        Y= start_date1
        X=(start_date1 + 6*day_delta)
        firstDate_time= Y.strftime("%Y-%m-%dT%H:%M:%S")
        LastDate_time= X.strftime("%Y-%m-%dT%H:%M:%S")
        print('FirstDate_time:',firstDate_time)
        print('LastDate_time:',LastDate_time)

        
        #mentioning the URl for data extraction
        url=''+param+'?startTime='+firstDate_time+'&endTime='+LastDate_time+'&tagId='+con_tagid
        r=requests.post(url,headers=API_KEY)
        data = r.json()
       
        #Normalizing the data
        works_data=json_normalize(data=data['data'],record_path=['telData'])
        
        #appending the data into one dataframe
        df1=pd.DataFrame.from_dict(works_data)
        dataframeJerk=pd.concat([df1,dataframeJerk],ignore_index=True)
        print(dataframeJerk)
        
        start_date1 += delta

    index = index + 1

dataframeJerk.columns

dataframeJerk['tag_id'].value_counts()

dataframeJerk.to_excel('Monthly_Rawdata_Jerk.xlsx',index=False)

# 1.4 Data Extraction- Monthly-Location

# Code for Extraction of Data through API

# The size of each step in days
dataframe=pd.DataFrame()
day_delta = timedelta(days=1)
delta = timedelta(days=7)
print(day_delta)
print(delta)

start_date=date

tagid=[2031010108,2031010506,2031010170]

index= 0

while index < len(tagid):
    con_tagid = str(tagid[index])
    print(con_tagid)
    
    Location=str('locations') 
    #print(Location)
    
    start_date1=start_date
    
    end_date = start_date1 + 31*day_delta
    print(end_date)
    Dategap=((end_date - start_date1).days)/7
    print("Dategap:",Dategap)


    for i in range(int(Dategap)):
        print ("Loop:",i)
        Y= start_date1
        X= (start_date1 + 6*day_delta)
        firstDate_time= Y.strftime("%Y-%m-%dT%H:%M:%S")
        LastDate_time= X.strftime("%Y-%m-%dT%H:%M:%S")
        print('FirstDate_time:',firstDate_time)
        print('LastDate_time:',LastDate_time)

        
        url=''+Location+'?startTime='+firstDate_time+'&endTime='+LastDate_time+'&tagId='+con_tagid
        r=requests.post(url,headers=API_KEY)
        data = r.json()
        data
        
        try:
            works_data=json_normalize(data=data['data'],record_path=['telData'])
        
            df=pd.DataFrame.from_dict(works_data)
            dataframe=pd.concat([df,dataframe],ignore_index=True)
            print(dataframe)
        except NotImplementedError as e:
            print(e)
        
        start_date1 += delta
    
    index = index + 1

dataframe.columns

dataframe['gw'].value_counts()

dataframe.to_excel('Monthly_Rawdata_Location.xlsx',index=False)
