==> 2. Cooler Box Converting Time Stamp for the Parameters.ipynb <==
import pandas as pd
import datetime

# 2.1 TS conversion- Monthly-Parameter- Light

Light=pd.read_excel('Monthly_Rawdata_Light.xlsx')

Light.head(5)

# convert ts to proper date format
s_time = pd.to_datetime(Light['ts']).dt.tz_localize(None)
s_time

Light['con_datetime']=s_time

Light.head()

# Set con_datetime as index
Light.set_index('con_datetime',inplace=True)
Light.head()

Light.drop(['trigger','ts'], axis = 1,inplace = True)

Light.head()

Light.to_excel('TS converted Light Raw Data.xlsx')

# 2.2 TS conversion- Monthly-Parameter- Temp

Temp=pd.read_excel('Monthly_Rawdata_Temp.xlsx')
Temp.head()

# convert ts to proper date format
s_time = pd.to_datetime(Temp['ts']).dt.tz_localize(None)
s_time

Temp['con_datetime']=s_time

Temp.head()

# Set con_datetime as index
Temp.set_index('con_datetime',inplace=True)
Temp.head()

Temp.drop(['ts'], axis = 1,inplace = True)

Temp.head()

Temp.to_excel('TS converted Temp Raw Data.xlsx')

# 2.3 TS conversion- Monthly-Parameter- Shock

Shock=pd.read_excel('Monthly_Rawdata_Jerk.xlsx')

Shock.head()

# convert ts to proper date format
s_time = pd.to_datetime(Shock['ts']).dt.tz_localize(None)
s_time

Shock['con_datetime']=s_time

Shock.head()

# Set con_datetime as index
Shock.set_index('con_datetime',inplace=True)
Shock.head()

Shock.drop(['ts'], axis = 1,inplace = True)

Shock.head()

Shock.to_excel('TS converted Jerk Raw Data.xlsx')

# 2.4 TS conversion- Monthly-Parameter- Location

location=pd.read_excel('Monthly_Rawdata_Location.xlsx')

location.head()

# convert ts to proper date format
s_time = pd.to_datetime(location['ts']).dt.tz_localize(None)
s_time

location['con_datetime']=s_time

location.head()

# Set con_datetime as index
location.set_index('con_datetime',inplace=True)
location.head()

location.drop(['ts'], axis = 1,inplace = True)

location.head()

location.to_excel('TS converted Location Raw Data.xlsx')
