==> 3. Cooler Box Merging Location data - Location vs Parameters.ipynb <==
# 3.1 Merging Temp data with location

import pandas as pd
import numpy as np
import datetime

Temp = pd.read_excel('TS converted Temp Raw Data.xlsx')

Temp.head()

Loc = pd.read_excel('TS converted Location Raw Data.xlsx')

Loc.head()

Loc['DateTime']=Loc['con_datetime'].dt.round('10min')

Loc.head()

Temp['DateTime']=Temp['con_datetime'].dt.round('10min')

Temp.head()

# Extract month & year from date
Temp['Month'] = Temp['DateTime'].dt.month
Temp['Year'] = Temp['DateTime'].dt.year

Temp.head()

# Add truck no column to identify data later
def Box(tag_id):
    if tag_id == 2031010108:
        return "AHB-003"
    elif tag_id == 2031010170:
        return "AHB-001"
    elif tag_id == 2031010506:
        return "AHB-002"
    else:
        return "NaN"

Temp["Box no"] = Temp["tag_id"].apply(lambda tag_id: Box(tag_id))

Temp.head()

# Add truck no column to identify data later
def Box(tag_id):
    if tag_id == 2031010108:
        return "AHB-003"
    elif tag_id == 2031010170:
        return "AHB-001"
    elif tag_id == 2031010506:
        return "AHB-002"
    else:
        return "NaN"

Loc["Box no"] = Loc["tag_id"].apply(lambda tag_id: Box(tag_id))

Loc.head()

lat_lon=Loc.copy()

# Add x column to concat con_datetime & Truck_no in temp & lat_lon dataframe

lat_lon['DateTime']=lat_lon['DateTime'].astype(str)

lat_lon['x']=lat_lon['DateTime']+lat_lon['Box no']
lat_lon

Temp['DateTime']=Temp['DateTime'].astype(str)

Temp['x']=Temp['DateTime']+Temp['Box no']
Temp

Temp['Box no'].value_counts()

# Performing Left join to find Location of Temp Data

Left_join = pd.merge( lat_lon, Temp,
                     on ='x', 
                     how ='left')
Left_join

Temp_Loc= Left_join.copy()

Temp_Loc['Box no_x'].value_counts()

Temp_Loc.dropna(axis=0, how='any', thresh=None, subset=None, inplace=True)

Temp_Loc.head()

Temp_Loc['Box no_y'].value_counts()

Temp_Loc.drop(['x','con_datetime_y','Box no_y'],axis=1,inplace=True)

Temp_Loc.head()

Temp_Loc.to_excel('Temp with Location.xlsx',index=False)

# 3.2 Merging Light data with location

Light = pd.read_excel('TS converted Light Raw Data.xlsx')

Light.head()

# Extract month & year from date
Light['Month'] = Light['con_datetime'].dt.month
Light['Year'] = Light['con_datetime'].dt.year

Light.head()

Loc = pd.read_excel('TS converted Location Raw Data.xlsx')

Loc.head()

Loc['DateTime']=Loc['con_datetime'].dt.round('10min')

Loc.head()

# Add truck no column to identify data later
def Box(tag_id):
    if tag_id == 2031010108:
        return "AHB-003"
    elif tag_id == 2031010170:
        return "AHB-001"
    elif tag_id == 2031010506:
        return "AHB-002"
    else:
        return "NaN"

Light["Box no"] = Light["tag_id"].apply(lambda tag_id: Box(tag_id))

Light.head()

# Add truck no column to identify data later
def Box(tag_id):
    if tag_id == 2031010108:
        return "AHB-003"
    elif tag_id == 2031010170:
        return "AHB-001"
    elif tag_id == 2031010506:
        return "AHB-002"
    else:
        return "NaN"

Loc["Box no"] = Loc["tag_id"].apply(lambda tag_id: Box(tag_id))

Loc.head()

lat_lon=Loc.copy()

lat_lon['DateTime']=lat_lon['DateTime'].astype(str)

lat_lon['x']=lat_lon['DateTime']+lat_lon['Box no']
lat_lon

Light['con_datetime']=Light['con_datetime'].astype(str)

Light['x']=Light['con_datetime']+Light['Box no']
Light

Light['Box no'].value_counts()

Left_join = pd.merge(lat_lon, 
                     Light, 
                     on ='x', 
                     how ='left')
Left_join

Light_Loc=Left_join.copy()

Light_Loc.head()

Light_Loc.dropna(axis=0, how='any', thresh=None, subset=None, inplace=True)

Light_Loc.head()

# Add column lidopen, where intensity >= 1,formula to calculate lidopen
Light_Loc['Lidopen']=np.where(Light_Loc['intensity']>=1,1,0)
Light_Loc.head()

Light_Loc.drop(['x','con_datetime_y','Box no_y','intensity'],axis=1,inplace=True)

Light_Loc.head()

# To display location where lidopen == 1
Result=Light_Loc.loc[(Light_Loc['Lidopen']==1)]
Result.head()

Result.to_excel('Lidopen with location.xlsx',index=False)

# 3.3 Merging Shock data with location

Jerk = pd.read_excel('TS converted Jerk Raw Data.xlsx')

Jerk.head()

Loc = pd.read_excel('TS converted Location Raw Data.xlsx')

Loc.head()

Loc['DateTime']=Loc['con_datetime'].dt.round('10min')

Loc.head()

# Add truck no column to identify data later
def Box(tag_id):
    if tag_id == 2031010108:
        return "AHB-003"
    elif tag_id == 2031010170:
        return "AHB-001"
    elif tag_id == 2031010506:
        return "AHB-002"
    else:
        return "NaN"

Jerk["Box no"] = Jerk["tag_id"].apply(lambda tag_id: Box(tag_id))

Jerk.head()

# Add truck no column to identify data later
def Box(tag_id):
    if tag_id == 2031010108:
        return "AHB-003"
    elif tag_id == 2031010170:
        return "AHB-001"
    elif tag_id == 2031010506:
        return "AHB-002"
    else:
        return "NaN"

Loc["Box no"] = Loc["tag_id"].apply(lambda tag_id: Box(tag_id))

Loc.head()

lat_lon=Loc.copy()

lat_lon['DateTime']=lat_lon['DateTime'].astype(str)

lat_lon['x']=lat_lon['DateTime']+lat_lon['Box no']
lat_lon

Jerk['con_datetime']=Jerk['con_datetime'].astype(str)

Jerk['x']=Jerk['con_datetime']+Jerk['Box no']
Jerk

Jerk['Box no'].value_counts()

Left_join = pd.merge(lat_lon, 
                     Jerk, 
                     on ='x', 
                     how ='left')
Left_join

Jerk_Loc=Left_join.copy()

Jerk_Loc.head()

Jerk_Loc.dropna(axis=0, how='any', thresh=None, subset=None, inplace=True)

Jerk_Loc.head()

# Add column lidopen, where intensity >= 1,formula to calculate lidopen
Jerk_Loc['Shock']=((Jerk_Loc['ampx']**2)+(Jerk_Loc['ampy']**2)+(Jerk_Loc['ampz']**2))**0.5

Jerk_Loc.head()

Jerk_Loc.drop(['x','con_datetime_y','Box no_y','ampx','ampy','ampz'],axis=1,inplace=True)

Jerk_Loc.head()

Jerk_Loc.to_excel('Shock with location.xlsx',index=False)
