==> 4. Cooler Box Reverse Geocoding - Parameters vs Location.ipynb <==
import pandas as pd
import numpy as np
import geopy
from geopy.geocoders import Nominatim
from geopy.extra.rate_limiter import RateLimiter
import tqdm
from tqdm.notebook import tqdm_notebook

# 4.1 Reverse Geocoding to Extract Address of Temp with Location

Temp=pd.read_excel('Temp with Location.xlsx')
Temp.head()

Temp['Geom'] = Temp['lat'].map(str) + ',' + Temp['long'].map(str)
Temp.head()

locator = Nominatim(user_agent='myGeocoder', timeout=10)
rgeocode = RateLimiter(locator.reverse, min_delay_seconds=0.001)

tqdm_notebook.pandas()
Temp['Address'] = Temp['Geom'].progress_apply(rgeocode)
Temp.head()

Temp.to_excel('Address of Temp with Location.xlsx',index=False)

# 4.2 Reverse Geocoding to Extract Address of Light with Location

Light=pd.read_excel('Lidopen with location.xlsx')
Light.head()

Light['Geom'] = Light['lat'].map(str) + ',' + Light['long'].map(str)
Light.head()

locator = Nominatim(user_agent='myGeocoder', timeout=10)
rgeocode = RateLimiter(locator.reverse, min_delay_seconds=0.001)

tqdm_notebook.pandas()
Light['Address'] = Light['Geom'].progress_apply(rgeocode)
Light.head()

Light.to_excel('Address of Light with Location.xlsx',index=False)

# 4.3 Reverse Geocoding to Extract Address of Shock with Location

Shock=pd.read_excel('Shock with location.xlsx')
Shock.head()

Shock['Geom'] = Shock['lat'].map(str) + ',' + Shock['long'].map(str)
Shock.head()

locator = Nominatim(user_agent='myGeocoder', timeout=10)
rgeocode = RateLimiter(locator.reverse, min_delay_seconds=0.001)

tqdm_notebook.pandas()
Shock['Address'] = Shock['Geom'].progress_apply(rgeocode)
Shock.head()

#conditions = [(Shock['Shock'] >= 0) & (Shock['Shock'] <=2),
#              (Shock['Shock']>= 2) & (Shock['Shock'] <= 6),
#              (Shock['Shock'] >= 6) & (Shock['Shock'] <= 100)]

#values = ['0-2','2-6', '>6']

#Shock['Condition'] = np.select(conditions, values)

#Shock.head()

Shock.to_excel('Address of Shock with Location.xlsx',index=False)
